Combine gnmi_cli utilities into a single file
- Combined client_parse_key_value and client_strzcpy into
  gnmi_cli_utils.

- Renamed top-level project to networking-recipe.

- Fixed line endings in README.md.
